2018-12-02 Implementando DAO

+++ ImplicareCommon/src/br/cefetmg/implicare/model/service/CargoManagement.java
@@ -11,7 +11,7 @@ public interface CargoManagement extends Remote{
 
     public ArrayList<Cargo> listar() throws PersistenceException, RemoteException;
     
-    public ArrayList<Cargo> listarCargoAreaEstudo(long cpf) throws PersistenceException, RemoteException;
+    public ArrayList<Cargo> listarCargoAreaEstudo(int cpf) throws PersistenceException, RemoteException;
 
     public Cargo pesquisar(int codCargo) throws PersistenceException, RemoteException;
 }
