+++ ImplicareCore/src/br/cefetmg/implicare/model/serviceImpl/CargoManagementImpl.java
@@ -22,7 +22,7 @@ public class CargoManagementImpl extends UnicastRemoteObject implements CargoMan
     }
 
     @Override
-    public ArrayList<Cargo> listarCargoAreaEstudo(long cpf) throws PersistenceException, RemoteException {
+    public ArrayList<Cargo> listarCargoAreaEstudo(int cpf) throws PersistenceException, RemoteException {
         return dao.listarCargoAreaEstudo(cpf);
     }
